[ refactor ] LabelContent 元件拆分
